+++ src/main/java/com/example/demo/controller/TransportController.java
@@ -2,13 +2,16 @@ package com.example.demo.controller;
 
 import com.example.demo.dto.ApiResponse;
 import com.example.demo.model.Transport;
+import com.example.demo.service.AuthService;
 import com.example.demo.service.TransportService;
 import com.example.demo.util.JwtUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/api/transports")
@@ -21,15 +24,29 @@ public class TransportController {
     @Autowired
     private JwtUtil jwtUtil;
 
+    @Autowired
+    private AuthService authService;
+
     // Add new transport
     @PostMapping("/")
-    public ResponseEntity<ApiResponse<Integer>> addTransport(
+    public ResponseEntity<?> addTransport(
             @RequestHeader("Authorization") String authHeader,
             @RequestBody Transport transport) {
         try {
             String token = authHeader.substring(7);
             Integer loggedInVendorId = jwtUtil.getUserIdFromToken(token);
             String userType = jwtUtil.getUserTypeFromToken(token);
+            if (!Objects.equals(userType, "VENDOR")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Only vendors are allowed to create transport.");
+            }
+
+
+            String serviceType = authService.getVendorServiceType(loggedInVendorId);
+            if(!Objects.equals(serviceType, "Transport_Provider")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Only Transport_Providers are allowed to create transports.");
+            }
 
             Integer transportId = transportService.addTransport(transport, loggedInVendorId);
             return ResponseEntity.ok(new ApiResponse<>(true, "Transport added successfully", transportId));
@@ -40,12 +57,22 @@ public class TransportController {
 
     // Get all transports of a vendor
     @GetMapping("/")
-    public ResponseEntity<ApiResponse<List<Transport>>> getTransportsByVendor(
+    public ResponseEntity<?> getTransportsByVendor(
             @RequestHeader("Authorization") String authHeader) {
         try {
             String token = authHeader.substring(7);
             Integer loggedInVendorId = jwtUtil.getUserIdFromToken(token);
             String userType = jwtUtil.getUserTypeFromToken(token);
+
+            if (!Objects.equals(userType, "VENDOR")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Only vendors are allowed to get transport.");
+            }
+            String serviceType = authService.getVendorServiceType(loggedInVendorId);
+            if(!Objects.equals(serviceType, "Transport_Provider")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Only Transport_Providers are allowed to get transports");
+            }
 
             List<Transport> transports = transportService.getTransportsByVendor(loggedInVendorId);
             return ResponseEntity.ok(new ApiResponse<>(true, "Transports fetched successfully", transports));
@@ -56,13 +83,24 @@ public class TransportController {
 
 
     @GetMapping("/{transportId}")
-    public ResponseEntity<ApiResponse<Transport>> getTransportByVendorAndId(
+    public ResponseEntity<?> getTransportByVendorAndId(
             @PathVariable Integer transportId,
             @RequestHeader("Authorization") String authHeader) {
         try {
             String token = authHeader.substring(7);
             Integer loggedInVendorId = jwtUtil.getUserIdFromToken(token);
             String userType = jwtUtil.getUserTypeFromToken(token);
+
+            if (!Objects.equals(userType, "VENDOR")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Only vendors are allowed to get transport by id");
+            }
+
+            String serviceType = authService.getVendorServiceType(loggedInVendorId);
+            if(!Objects.equals(serviceType, "Transport_Provider")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Only Transport_Providers are allowed to get transport by id.");
+            }
 
             Transport transport = transportService.getTransportByVendorAndId(loggedInVendorId, transportId);
             return ResponseEntity.ok(new ApiResponse<>(true, "Transport fetched successfully", transport));
@@ -73,7 +111,7 @@ public class TransportController {
 
     // Update transport
     @PutMapping("/{transportId}")
-    public ResponseEntity<ApiResponse<Integer>> updateTransport(
+    public ResponseEntity<?> updateTransport(
             @PathVariable Integer transportId,
             @RequestHeader("Authorization") String authHeader,
             @RequestBody Transport transport) {
@@ -81,6 +119,17 @@ public class TransportController {
             String token = authHeader.substring(7);
             Integer loggedInVendorId = jwtUtil.getUserIdFromToken(token);
             String userType = jwtUtil.getUserTypeFromToken(token);
+
+            if (!Objects.equals(userType, "VENDOR")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Only vendors are allowed to update transport.");
+            }
+
+            String serviceType = authService.getVendorServiceType(loggedInVendorId);
+            if(!Objects.equals(serviceType, "Transport_Provider")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Only Transport_Providers are allowed to update transports.");
+            }
 
             int updated = transportService.updateTransport(loggedInVendorId, transportId, transport);
             return ResponseEntity.ok(new ApiResponse<>(true, "Transport updated successfully", updated));
@@ -91,13 +140,24 @@ public class TransportController {
 
     // Delete transport
     @DeleteMapping("/{transportId}")
-    public ResponseEntity<ApiResponse<Integer>> deleteTransport(
+    public ResponseEntity<?> deleteTransport(
             @PathVariable Integer transportId,
             @RequestHeader("Authorization") String authHeader) {
         try {
             String token = authHeader.substring(7);
             Integer loggedInVendorId = jwtUtil.getUserIdFromToken(token);
             String userType = jwtUtil.getUserTypeFromToken(token);
+
+            if (!Objects.equals(userType, "VENDOR")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Only vendors are allowed to delete transport.");
+            }
+
+            String serviceType = authService.getVendorServiceType(loggedInVendorId);
+            if(!Objects.equals(serviceType, "Transport_Provider")) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Only Transport_Providers are allowed to delete transports.");
+            }
 
             int deleted = transportService.deleteTransport(loggedInVendorId, transportId);
             return ResponseEntity.ok(new ApiResponse<>(true, "Transport deleted successfully", deleted));
